fix(tests): refactor blueprints into seperate files for each blueprint type to reduce file sizes
