+++ tests/conftest.py
@@ -5,7 +5,6 @@ import os
 import shutil
 import socket
 import sys
-import uuid
 from contextlib import AsyncExitStack
 from datetime import datetime, timezone
 from typing import Any, AsyncGenerator, Callable, Generator
